extract selectapplicationlist into a custom component
